Make Base Table Component Reusable/Modular
All main list views (Employee, Hardware & Finance) have many things in common,
the only varying datas being: the endpoint, the columns & the actual row values
(properties of the API response object).

The 3 varying data can be passed as props to the base table for each list
view: the enpoint as a string, the columns as an array of strings and the
row values as an array of properties of the object (as string).

This reduces repetitive code and improves performance of the application.